requirements.txt

viam-sdk==0.58.0

typing-extensions
speech_service_api @ git+https://github.com/viam-labs/speech-service-api.git@v0.6.0
